Fix cannot be narrowed to type 'uintptrj_t' error
A lot of the following compilation errors occur on OMROptions.cpp during
compilation with Clang 8:

..\..\compiler\control\OMROptions.cpp(760,94):  error: constant expression evaluates to -2147483648 which cannot be narrowed to type 'uintptrj_t' (aka 'unsigned long long') [-Wc++11-narrowing]
{"enableTraps", "C\tenable trap instructions", RESET_OPTION_BIT(TR_DisableTraps), "F"},

The reason is a narrowing the latest component of the RESET_OPTION_BIT
macro '((x)&~TR_OWM)' to the 'uintptrj_t' aka
'unsigned long long' type. The expression must be explicitly converted
to 'uintptrj_t'.

Closes: #3121
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>