Add asserts validating correct APIs are used for RX,RS,SI instructions
Add asserts for long displacement `generate*Instruction` APIs to make
sure we've used the correct API. This helps prevent situations where in
the past we may have used `generateRXInstruction` for example to create
an RXY instruction and the `useSourceMemoryReference` API would then
"upgrade" the RX instruction into an RXY instruction by changing it's
instruction kind.

Since instruction kind is now immutable this is no longer possible and
hence we would be encoding an RXY instruction as an RX instruction
which is incorrect at the binary level.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>