Map generateRXYInstruction APIs to generateRXInstruction APIs
After examining the number of locations that must be modified so proper
`generateRX*` API variants are used for RX and RXY instructions the
change does not seem feasible. It can be done of course but at the cost
of code cleanliness as there are many locations where the actual
mnemonic that will be used is not known until runtime.

In such cases we would have to effectively check the mnemonic format at
every such location and determine whether we need to generate an RX or
RXY instruction.

Instead of doing this we effectively deprecate the RXY API variants and
handle both RX and RXY instructions in the `generateRXInstruction` API.
This simplifies generating RX-type instructions into a common place and
guarantees we will do the right thing. We'll have follow-up commits to
handle RS and SI instructions as well.

Once downstream dependencies have been removed we'll deprecate the APIs.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>