Don't use `generateRXInstruction` API for non-RX(Y) instructions
We must ensure that we do not use the `generateRXInstruction` APIs for
non-RX(Y) instructions as we have now placed asserts to make sure we
are generating instructions of the proper type.

A typical example of where we get this wrong today is when generating
RXE instructions which most of the time do not have a mask and behave
more or less like 6-byte RX instructions.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>