Fix result loading from unresolved class data snippet
An unresolved class is resolved at runtime by a PicBuilder routine and the
result (resolved class) is stored in a patchable literal pool
slot referenced by the last slot of an 'unresolved data snippet'.

On Z hardware lower than z10, the JIT compiled code does not use the
correct slot of the snippet and, as a result of this, crashes when loading
the resolved class from the data snippet. This bug is fixed by this
commit.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>