Only count upgraded long displacement static debug counters
When "upgrading" a short displacement to a long displacement
instruction during binary encoding only count the static debug counter
if the upgrade actually occurred. The same path is taken for upgradable
instructions as is with instructions which are already of long
displacement format (RXY, RSY, SIY) so we need an extra check to avoid
false positives.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>