Add CODEOWNERS file for automated review requests
As per the discussion in [1] and the subsequent discussion within the
PR itself we add a CODEOWNERS file at the top-level and specify the
code owners for particular areas within the Eclipse OMR project.

The owners of the corresponding areas will automatically get review
requests when a PR is opened touching files within their respective
area of expertise.

@charliegracie and @mstoodle are identified as global catch-all code
owners for general contributions to the repository.

[skip ci]

[1] https://eclipse-omr.slack.com/archives/CCQ8B4B39/p1541599655022400

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>