Guarantee functional correctness of rdbar/wrtbar
Map rdbar/wrtbars to corresponding opcode to normal opcode to keep
consistency with the existing wrtbar opcode. Subprojects can overwrite
the mapping decisions.

PRE and FieldPrivatizer are disabled for now if the trees might contain
newly added rdbar/wrtbar. All the transformations applied to normal
load/write s should be applied to rd/wrtbar s. However, we need to
be careful about the difference in the shape of the trees and the
mapping relationship between different loads and writes. More
investigation is needed to make sure those 2 opts would do the right
thing for rdbar/wrtbars.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>