Clean up CMake defines
Rename OMR_JIT to OMR_COMPILER in CMake. Since OMR_JIT also existed
in the auto-tools build system change it there as well.

Clean up the AppVeyor.cmake and Travis.cmake CMake caches to remove
variables being set to their default value, remove variables that
are never used and group variables by component.

In the cmake.config group the major flags by "component" and
modify the default value for flags to be more appropriate. If either
of the components which build compiler targets are enabled verify
that the compiler component is enabled. If it is not error out.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>