Allows iterator to complete a cycle irrespective of file availability
Its observed that the iterator is going to infinity loop if it doesn't
find the cgroup metric file in the /sys/fs/cgroup/<subsystem>/<file>
so added changes to skip and proceed if a file is not available so that
other metric displaying task is uninterrupted.

Added a macro (CGROUP_METRIC_FILE_CONTENT_MAX_LIMIT) to define the max
limit of cgroup metric file content. Removed duplicate documentation

Signed-off-by: bharathappali <bharath.appali@gmail.com>