AArch64: Fix createPrologue()
createPrologue() has saved the argument registers using strimmx
or vstrimmd instructions, regardless of the data size.  On the other
hand, the stack offset for the argument registers takes the data
size into consideration.
This commit changes the createPrologue() instruction generation, so
that 32-bit and smaller arguments are saved by strimmw or vstrimms
instructions.

Signed-off-by: knn-k <konno@jp.ibm.com>