Use an int to store the output of ioctl FIONREAD
All publicly available documentation indicates that FIONREAD outputs an
int, not a pointer-width value. Use an int, and avoid bad aliasing.
There is a comment indicating that _some_ devices output a 64 bit value,
but I cannot see how or when that would ever be true. This PR removes
that comment.

Fixes: #3145
Signed-off-by: Robert Young <rwy0717@gmail.com>