Deprecate Z codegen `startOfLiveRange` and unify ICF handling on Z
The `startOfLiveRange` function was used by true register model to
identify the start of the live range of a particular register. With the
deprecation of true register model this is no longer necessary.

To calculate this value special care was needed around ICF regions. As
such expensive tracking was needed for all instructions within an ICF
regions and it involved an O(n^2) algorithm to collect the right
information as well as storing it within a dynamic list.

With this removal we save quite a few compile time cycles which makes
our register allocator faster.

We take this opportunity to unify the Z codegen with the common APIs
for keeping track of nested ICF regions as is already implemented on all
the other platforms using common APIs. We mark the common fields in the
base code generator private as part of this change to prevent future
uses from using the field directly as they must interact using the
publicly available APIs.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>