Add new helper induceOSRAtCurrentPCAndRecompile
Add the new helper induceOSRAtCurrentPCAndRecompile to request both OSR
transition and method reompilation. When OSR is requested
because of invalidated code path the method body containing
the invalid code path should be dicarded and recompiled for better
performance.

The old helper induceOSRAtCurrentPC is still useful if only OSR
transition is needed.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>