Fix error when array copy length is constant 256
When node is set to be backwardArrayCopy node, we generate backward array copy
sequence only. When this is combined with constant length bytelength we end up
with NULL byteLenReg which causes segmentation fault.
Also When constant byte length is exactly 256 we do not have
internal control flow so update the condition to not set end ICF in this case.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>