Add OmrOption, a util for declaring build options
OmrOption is a module which provides a declarative syntax for specifying
and validating ON/OFF build options. The long-term goal is to declare
and validate all OMR build options using this module.

The general syntax is:

  omr_add_option_group(<group>)

  omr_add_option(<option> <option-group>
    COMMENT <comment>
    DEFAULT <expression>
    REQUIRES <option>...
    REQUIRES_EXPR <expr>
    CONFLICTS <option>...
  )

  omr_validate_option_group(<group>)

Options are defined as BOOL cache variables.

DEFAULT is an expression, which is evaluated and coerced to an on/off
value. The default value is only evaluated on the first run, and cached
in subsequent runs. Since options are regular cmake cache variables, the
value can be manually overridden by users at any point. A default value
must always be provided.

REQUIRES allows users to pass a list of required flags or variables,
which must be ON if the option is ON.

Similarly, CONFLICTS is a list of options which must be OFF, if the
option is ON.

REQUIRES_EXPR is an expression which must be TRUE, if the option is
ON. This can be used for expressing more complex option requirements.
Each option can have only one associated REQUIRES_EXPR.

Options are registered into a group, which must be explicitly validated
with a call to omr_validate_option_group. Option groups allow consuming
projects to create and validate their own option sets. By making
validation explicit, we allow the user (us OMR developers) to delay
validation until after all configuration variables have been set.
Options are validated on every run of cmake, which ensures that the user
never puts OMR into an illegal configuration.

The phrasing "option" was chosen over "flag" to follow existing cmake
conventions, where add_option / add_dependent_option create ON/OFF
bool cache variables.

There is a similar module built in to cmake called DependentOption.
This module was not used because:
1. A dependent option is only show to the user if it's dependencies are
   satisfied. OmrOption makes the choice to always present ALL flags in
   the cache.
2. A dependent option will revert to it's default value if it's
   requirements aren't satisfied, overriding any user provided value.
   OmrOption never tries to override the user, and relies on option
   validation instead.
3. OmrOption provides an easier to read, more obvious syntax.
4. The default value in OmrOption is an expression, rather than a simple
   on/off constant. This gives us a succinct way to implement implicitly
   enabled/disabled flags, while still being user overridable.
   Dependent options have a dependency expression, but the default value
   is a constant, which cannot be overridden by the user.

Signed-off-by: Robert Young <rwy0717@gmail.com>