Simplify instanceof/guard type constraint creation
...where genTypeResult() was used.

Rather than attempt not to intersect constraints, this logic now simply
generates the constraint for objects that pass the type test, and tries
to add that constraint on the appropriate path, implicitly intersecting
it with any pre-existing constraint. The branch can be folded away if
one of the paths is found to be unreachable due to an empty constraint
intersection.

The previous logic could sometimes fold away a profiled guard against
java/lang/Class even in the absence of any other type information,
generating an unconditional goto to the taken side (fall-back path),
which is functionally correct but undesirable. This now only happens
when an existing type constraint precludes the object from being an
instance of java/lang/Class.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>