Add extra assertions for atomic non-helpers on Power
Added assertion that we're running a 64-bit JVM when trying to inline
one of the atomic non-helpers, as 32-bit code generation is not yet
supported for it.

Also if a call to codegen inlineDirectCall from directCallEvaluator does
not inline a call, added an assertion that the symbol is not one of the
atomic non-helpers.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>