Initialize Suballocator variables all platforms
Suballocator related variables suballocatorInitialSize and
suballocatorCommitSize in MM_GCExtensionsBase are defined for all
platforms but initialized in class constructor only for Compressed Refs.
It cause operations with non-initialized memory might crash Memory
Check. Extending scope of initialization of these variables for all
platforms. As far as Suballocator happen to be used for Compressed Refs
platforms only it is irrelevant which values these variables are
initialized on other platforms. Set them exactly the same as in CR to
have cleaner code. Again, this would not have any impact for non-CR
platforms footprint.

Signed-off-by: Dmitri Pivkine <Dmitri_Pivkine@ca.ibm.com>