Split JitBuilder API generator tests into multiple files
Instead of putting all the tests for the different components of the
generator in one file, it is much more convenient to have a separate
file for each major component. This commit therefore splits the
existing tests from `jitbuilder/apigen/test.py` into:

- `jitbuilder/apigen/test/apidescriptiontests.py`
- `jitbuilder/apigen/test/genutilstests.py`

A `__init__.py` file is also added to allow the tests file to "reach up"
and access the generator implementation modules.

Lastly a `run_tests.py` file is added to actually execute all the tests
and the corresponding ctest command is updated to execute this new file.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>