Deprecate displacement field inside of S390 MemoryReference
The offset and displacement fields inside of S390 memory references are
very confusing. The offset field tracks the offset relative to whatever
entity is being referenced, however the true displacement which we will
encode for the memory reference may not be known until binary encoding
time because the memory reference could be referencing stack symbols or
snippets which have not been mapped up until binary encoding.

In this commit we deprecate the displacement field along with its
corresponding public query and we instead update the offset field to
contain the true displacement after binary encoding of the memory
reference.

Not that because S390 memory references can only be used once we do not
risk "re-updating" the offset because we will never call
`generateBinaryEncoding` twice on the same memory reference.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>