Restrict setting TR_DisableInternalPointer to current compilation
Previously, this option was set only when OMR::Compilation::compile()
was called. That resulted in the option not having an effect on
CodeGenerator's _disableInternalPointers member as that is
initialized much earlier in its constructor by checking a few
compilation-specific options, including TR_DisableInternalPointers.

For this reason, not setting TR_DisableInternalPointer option
globally was resulting in performance regression in some cases.
Setting this option  globally is not the correct thing to do,
as this causes the option to be applied to all subsequent
compilations regardless of whether it should be set.

By setting TR_DisableInternalPointer to the current
compilation-specific options in the the constructor of
OMR::Compilation before allocateCodeGenerator(TR::Compilation)
is called, we no longer need to set this option globally.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>