Add OMR_GC_TEST compile flag
Allow us to enable the GC, and FV testing, while turning the xml-driven
GC heap tests off. These tests are currently incompatible with the
gc-api and om.

In the future, the tests will be made to be gc-api compatible, but it's
unclear if the tests will be based off om or continue to use the object
model from the gc example.

Signed-off-by: Robert Young <rwy0717@gmail.com>