Refactor C++ generator into a class
Putting the C++ API generator functions in a class improves
encapsulation and allows overriding, which will be useful when
implementing other generators.

The key change in this commit, after putting the functions in a
class, is to add the `self` argument to all functions. Some minor
cleanup is also done.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>