Make x86-64 table jump position independent in AOT
The TR_MethodAddress relocation which has until now been generated for
the displacement of the memory reference of the jump dispatching a table
node is incorrect on 64-bit.

First, the displacement is only 32 bits, but the relocation works on a
pointer-sized value (64 bits). The high half of the "address" is really
the first four bytes of the next instruction. If the address adjustment
changes the high half, then the next instruction will be modified.

Second, the 32-bit displacement gets sign-extended. If the code were
originally compiled at a location below 2GB, and later loaded above 2GB
(but below 4GB), the address adjustment will set the sign bit, so that
the sign-extension is no longer the same as the zero-extension (i.e. the
correct address).

To prevent these problems, the table evaluator now generates a
position-independent sequence for the jump table dispatch, which
requires no (external) relocation:

    lea rJumpTable, [rip + OFFSET_TO_JUMP_TABLE]
    jmp qword ptr [rJumpTable + 8*rIndex]

This sequence is only generated for 64-bit relocatable code. 32-bit code
doesn't have access to rip-relative addressing, but it's immune to the
above relocation problems anyway, since the pointer size is 32 bits, and
there is no sign- nor zero-extension. For non-relocatable code, the
existing lone jmp is preferable to a two-instruction sequence.

The addresses contained in the jump table remain position-dependent, and
they continue to get relocated just as before.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>