Fix warnings in multiple (mostly compiler/optimizer)files
Newer compilers complain about a few things that are not syntactically
incorrect, but could be bugs. For example, not wrapping && expressions
in parenthesis when used in || conditionals. This commit fixes a number
of such warnings found primarily in the compiler/optimizer files. The
primary benefit of this cleanup is fewer distracting warning messages
during builds.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>