Provide Dockerfile for easy x86 build failure reproduction
While working on a recent PR on my MacBook, I needed to reproduce
a failure in the Linux x86 testing. To do that, I built a Docker
container to build and run tests and to diagnose the failure.

It seemed like something others might find useful, so I cleaned
it up a bit to make it a bit more consumable. It's not perfect
but its gets you to having a workin Linux x86 build with only
a couple of commands.  The basic directions are provided in
the Dockerfile, which I'll reproduce here to save a bit of time:

Basic usage (adapt to your environment and scenario):
  git clone https://github.com/eclipse/omr (or cd to a directory that contains an OMR repository)
  cd omr/buildenv/docker && docker build -f x86_64/ubuntu16/Dockerfile -t buildomr .
  cd ../../../
  docker run --privileged -v $PWD/omr:/omr -it buildomr

The container will automatically run cmake to configure itself and
then initiate the build.  Once the build completes, you'll be left
sitting at a prompt in the container so you can run more commands.
I had the container also install gdb, vim, and emacs but we could
add some other useful commands too.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>