Supporting options and code to allow AOT at cheap warm level
This commit adds a new option -Xaot:disableAOTAtCheapWarm to control the
logic that will force AOT compilations to be performed at a "cheap"
warm level. A "cheap" warm level is an optimization level where the
inliner has been subdued somewhat (the same inliner changes are applied
to the Xtune:virtualized mode in OpenJ9 JVM).
Upgrades of AOT bodies through sampling are also subdued if the
guarded counting recompilation (GCR) is enabled.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>