Perform wrtbar versioning conservatively
If a GC point is triggered mid-loop, destination object of
the wrtbar can move. And if it is not remembered, upon return
from GC the remaining loop iterations will not run any wrtbars
since we had versioned them out. The fix here is to be conservative
when versioning wrtbars - for correctness, if there is a GC
point in the loop, wrtbar versioning will not be performed.

Signed-off-by: Yan Luo <Yan_Luo@ca.ibm.com>