Implement switchCodeCacheTo() in CodeGenerator
Migrate the FrontEnd function `switchCodeCache` to the CodeGenerator where it
more naturally belongs, and rename it to `switchCodeCacheTo`.

There is another implementation in Compilation that is called from the FrontEnd
version that a number of places in the OMR codebase and downstream projects call.
However, each of those places should really be calling the FrontEnd version
instead as they also implement this missing functionality inline that the FrontEnd
version provides.

This commit introduces `switchCodeCacheTo` as the new API that will be used
throughout.  However, downstream projects will require modification so the existing
APIs remain intact until those changes can be made.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>