Add linkage tests for Microsoft calling convention
Microsoft x64 calling convention brings addition limitations for x86-64:
only four registers are available for arguments (not four for integer
and four for float but four for both). Also the caller has
responsibility to allocate 32 bytes of "shadow space" on the stack right
before calling the function (regardless of the actual parameters used).

Two tests are added to check these circumstances: the first just
demonstrates how to invoke a method with five arguments while the second
one invokes a method who is actively using the stack and can corrupts
the return address if no "shadow space" has been allocated.

Issue: #3366
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>