Add linkage tests with mixed arg types
Microsoft x64 calling convention brings addition limitations for x86-64:
only four registers are available for arguments (not four for integers
and four for floats but four for both). It's very interesting how the
compiler handles the situation when we need to call a function with four
mixed, float and integer, arguments and to call a function with more
than four arguments (the fifth and others must be placed onto the stack).

Issue: #3366
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>

Switch int to int32_t in custom defined functions

From @Leonardo2718:
By convention, we tend to use `int32_t` for integer types instead of
just `int` in the OMR compiler and related tests.

Signed-off-by: Pavel Samolysov <samolisov@gmail.com>