Deprecate TR_ForceLargeRAMoves
This option is no longer used as it aided debugging of the use of
64-bit registers on 31-bit since it forced (some) register shuffles and
spills to be 64-bit. The option did not cover all corners of RA however
so it is quite inconsistent and not too useful.

Rather than polluting already complicated logic with this half-baked
option we instead remove it completely.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>