Remove TR_GPR64 dependency from registerCopy and registerExchange
This requires checking whether the target virtual register is a 64-bit
register or not. We also add assertions to make sure the virtual
register is also available.

We note that there is very likely bugs in these APIs as they're
currently completely undocumented. For example it is non-obvious
whether the APIs can handle for example a register exchange of a 64-bit
and a 32-bit register.

We also correct a few bugs related to HPRs in the register exchange
API. It previously assumed that if both source and target registers are
not HPRs we will always use a 32-bit load, which is definitely not
correct because the virtual registers may be 64-bit.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>