Fix JitBuilder API description of VirtualMachineOperandStack constructors
It appears the constructor for VirtualMachineOperandArray was mistakenly
represented as VirtualMachineOperandStack constructor (fortunately has
a similar type signature if the last two optional parameters aren't used).
But the two final optional parameters aren't represented which means
the client API always forces the defaults (which obviously could be
wrong in certain scenarios).

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>