Most JitBuilder tests can (and should!) run on non-Windows x86 platforms
Many tests contain Calls that are currently not guaranteed to work
on other platforms until we fix our method trampoline problem.
But they should work on most X86 platforms and we should be running
these tests regularly in our PR tests on those x86 platforms to ensure
they are not broken (sadly, one of them was but that has been fixed now).
The extended tests do not currently work on Windows not because of
trampolines but due to lack of MSVC 2010 support for part of the
toIlType service in TypeDictionary.

At the same time, I renamed a somewhat confusingly named configuration
flag called OMR_JITBUILDER_ADDITIONAL to OMR_JITBUILDER_TEST_EXTENDED.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>