Optimize sequencial recognized object shadow field load in VP
Optimize the generalized pattern `Field1 f1 = o.field1; if (f1==null)
initialize_f1; Field2 f2 = f1.field2; if (f2==null) initialize_f2` where
`o` is a known object. If value of `f1` is null, it is initialized just
once to a valid non-null value and does not change; similarly, if `f2` is
null, it is also initialized to a valid non-null value and also does not
change during entire run time. If `o` is a known object. VP loads `f1` at
compile time and checks for null. A non-null value means it has been
initialized so we treat it as a known object and attach known object,
non-null, and fixed class constraints. In the loading of `f2`, the newly
propagated known object constraint is attached to the base object `f1`, so
again we do a compile-time load of `f2` and attach the appropriate
constraints. The "initialize-just-once-and-never-change" nature of the
field loads are represented as special `recognized known object` fields,
with the shadow sym refs specially tagged.

A Java example is `MyEnumClass.class.getEnumConstants()`. The method
retrieves the enum constants array via 2 indirect loads: EnumVars enumVars
= classObject.enumVars followed by Object[] enumConstants =
enumVars.cachedEnumConstants. Both fields are initialized once and never
change.

Signed-off-by: Yan Luo <Yan_Luo@ca.ibm.com>