Check and fix to make sure rdbars are anchored
This change adds check to make sure the rdbars are anchored
and anchor a rdbar if it's found unanchored.

Ideally all optimizations should anchor a rdbar when it's created.
This API is for fixing the unanchored rdbars in case some
optimization forget to anchor the node by mistake. Optimizations
should still try to generate correct rdbar trees instead of relying
on this API.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>