Copy OSR exception edge from caller when block is not split
Inliner unconditionally merges the last block of callee and the caller's
block immediately after and copies the caller block's non-OSR exception
successors to callee blocks. Then it splits the merged block when
needed. Split is not needed when no guard and no OSR blocks are
generated for the callee. However, the OSR exception successor from the
caller, must be copied onto the merged block, otherwise the OSR points
from the caller block will be missing OSR exception edges.

This commit does the following:
1. Copy the missing OSR exception edge
2. Ask OSR method data whether OSR blocks have been generated for the
   callee instead of checking the trees and crossed basic blocks
3. Improve the code readability by changing variables names and adding
   comments
   - Rename `blockOfCaller` to `mergedBlock`
   - Rename `blockOfCallerInCalleeCFG` to `blockOfTreesAfterTheCall`
   - A static function `succIsOSRCatchBlock` is added to only copy OSR
   exception successor

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>