Arraycopy opt needs to know about read barriers
There is a transformation that changes a reference arraycopy node into a
primitive arraycopy node if neither an arraystore check nor a write
barrier are required. This transformation should also not occur if a read
barrier is required.

transformArrayCopyCall was changed to check if read barriers are required
and elect not to transform the reference arraycopy node if this is the
case.

Issue: #3440
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>