Add helper function to free a TR::CodeCacheMemorySegment
The new function, `freeCodeCacheSegment()`, is intended to allow code
cache segments allocated by `allocateCodeCacheSegment()` to be freed.
Because it's not always safe to free code cache memory, the default
implementation does nothing. Callers should therefore treat the
function as a hint, since memory may not actually be freed.

A call to `freeCodeCacheSegment()` is also added to free the code cache
repository segment memory when the code cache manager is destroyed. An
important shortcoming of this change is that it does not try to free
related objects, such as the code cache repository object or code cache
repository monitor. However, the `TR::CodeCacheSegment()` instance
corresponding to the repository segment memory will be freed because
is embedded in the memory segment itself.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>